# MRG: Speed up browse figure creation if scalings are provided