This avoids loading data from disk or performing calculations if we don't actually need to calculate `scalings` parameters.

Also adds a `scalings` parameter to `Report.add_raw()`.


Performance improvements:

```
this branch: 2.17 s ± 214 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main:        4.07 s ± 625 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

with the following code:
```python
# %%
from pathlib import Path
import mne

sample_dir = Path(mne.datasets.sample.data_path())
sample_fname = sample_dir / 'MEG' / 'sample' / 'sample_audvis_raw.fif'

raw = mne.io.read_raw_fif(sample_fname, verbose=False)
# raw.pick([0])
raw.pick_channels(['MEG 0111', 'MEG 0112', 'EEG 001', 'ECG 001', 'STI 014'])
report = mne.Report()

# %%
%timeit report.add_raw(raw=raw, title='foo', psd=False, replace=True)
```

x-ref #10107